Optimize the addition of the linked list with a pointer to the last element for merge-k-sorted
There was an issue that when I was keeping a reference to the head of the result list and
I was traversing the whole list when adding an element. This is pointless and that why
I've optimized the code a bit so that I keep a pointer to the last element. This changed the
addition time to a O(1). The overall submission processing time dropped from 1680 ms to 7 ms!!!!